优化库存同步 更改库存金额字段类型

+++ src/main/java/com/aiqin/bms/scmp/api/purchase/domain/PurchaseOrderProduct.java
@@ -135,7 +135,7 @@ public class PurchaseOrderProduct {
 
     @ApiModelProperty(value="库存金额")
     @JsonProperty("stock_amount")
-    private Integer stockAmount;
+    private Long stockAmount;
 
     @ApiModelProperty(value="库存周转期")
     @JsonProperty("stock_turnover")
